Move most tempating logic to Template.
