+++ packages/shared-ui/src/elements/input/text-editor/text-editor.ts
@@ -9,69 +9,7 @@ import { createRef, ref, Ref } from "lit/directives/ref.js";
 import { Project } from "../../../state";
 import { FastAccessSelectEvent } from "../../../events/events";
 import { FastAccessMenu } from "../../elements";
-
-type TemplatePart = {
-  type: string;
-  path: string;
-  title: string;
-};
-
-type TemplatePartCallback = (part: TemplatePart) => string;
-
-class Template {
-  #renderableValue = "";
-
-  constructor(public readonly raw: string) {
-    raw = raw.trim();
-
-    this.#renderableValue = raw;
-    if (raw === "") {
-      this.#renderableValue = "&nbsp;";
-    }
-  }
-
-  get renderable() {
-    return this.#renderableValue;
-  }
-
-  substitute(callback: TemplatePartCallback) {
-    const value = this.raw;
-    const finder = /{{\s?(.*?)\s?\|\s?"(.*?)"\s?\|\s?"(.*?)"\s?}}/gim;
-    const matches = [];
-    let res;
-    do {
-      res = finder.exec(value);
-      if (res) {
-        matches.push(res);
-      }
-    } while (res);
-
-    if (matches.length === 0) {
-      return;
-    }
-
-    this.#renderableValue = "";
-    let current = 0;
-    for (const match of matches) {
-      const [str, type, path, title] = match;
-      if (current < match.index) {
-        this.#renderableValue += value.slice(current, match.index);
-      }
-
-      // To keep things a bit simpler in the regexp and so forth we send this
-      // out as a single line string.
-      this.#renderableValue += callback({ type, path, title });
-      current = match.index + str.length;
-    }
-
-    if (current < value.length) {
-      this.#renderableValue += value.slice(current);
-    } else {
-      // Ensure that if the final item is a chiclet we add a space on.
-      this.#renderableValue += "&nbsp;";
-    }
-  }
-}
+import { Template } from "../../../utils/template";
 
 @customElement("bb-text-editor")
 export class TextEditor extends LitElement {
@@ -300,8 +238,8 @@ export class TextEditor extends LitElement {
     label.classList.add(type);
     label.dataset.path = path;
 
-    preamableText.textContent = `{{ ${type} | "${path}" | "`;
-    postamableText.textContent = `" }}`;
+    preamableText.textContent = Template.preamble({ title, path, type });
+    postamableText.textContent = Template.postamble();
     titleText.textContent = title;
     titleText.classList.add("visible");
